Break up main nuxt module into sub modules
For easier maintainability
